requirements.txt
numpy==2.0.1
openai==1.40.6
pydub==0.25.1
python-dotenv==1.0.1
sounddevice==0.5.0
